+++ changelog.md
@@ -1,6 +1,8 @@
 ## Changelog
 
 ### v0.6.1
+* CHANGE | Separated mode select from main menu.
+* FIX | Removed unused font.
 * FIX | Preferences are now properly saved.
 * FIX | Fixed weird behaviour when loading diffrent profiles.